# Specialize map and map! for BitArrays