--- a/base/bitarray.jl
+++ b/base/bitarray.jl
@@ -1545,42 +1545,28 @@ maximum(B::BitArray) = isempty(B) ? error("argument must be non-empty") : any(B)
 
 ## map over bitarrays ##
 
-function map!(f::Callable, A::Union(StridedArray,BitArray))
+function map!(f::Callable, A::StridedArray)
     for i = 1:length(A)
         A[i] = f(A[i])
     end
     return A
 end
 
-function map!(f::Callable, dest::Union(StridedArray,BitArray), A::Union(StridedArray,BitArray))
+function map!(f::Callable, dest::StridedArray, A::StridedArray)
     for i = 1:length(A)
         dest[i] = f(A[i])
     end
     return dest
 end
 
-function map!(f::Callable, dest::Union(StridedArray,BitArray), A::Union(StridedArray,BitArray), B::Union(StridedArray,BitArray))
+function map!(f::Callable, dest::StridedArray, A::StridedArray, B::StridedArray)
     for i = 1:length(A)
         dest[i] = f(A[i], B[i])
     end
     return dest
 end
 
-function map!(f::Callable, dest::Union(StridedArray,BitArray), A::Union(StridedArray,BitArray), B::Number)
-    for i = 1:length(A)
-        dest[i] = f(A[i], B)
-    end
-    return dest
-end
-
-function map!(f::Callable, dest::Union(StridedArray,BitArray), A::Number, B::Union(StridedArray,BitArray))
-    for i = 1:length(B)
-        dest[i] = f(A, B[i])
-    end
-    return dest
-end
-
-function map!(f::Callable, dest::Union(StridedArray,BitArray), As::Union(StridedArray,BitArray)...)
+function map!(f::Callable, dest::StridedArray, As::StridedArray...)
     n = length(As[1])
     i = 1
     ith = a->a[i]