Alright, next step in getting IntSets to use BitArrays: a faster map.

Specializing map is even more important for bitarrays than it is for generic arrays since there can be a 64x+ speedup by working at the level of Int64 instead of looping bit-by-bit. There's no other way to perform operations at the chunk-level without either exporting more functions or mucking around with BitArray's internal fields (See mbauman/IntSets.jl#1).

This adds functor types (a la reduce.jl) that allow mapping across BitArrays 64-bits at a time. Performance of `map(&, x, y)` is now ~300x faster and very competitive with `x & y` for large arrays. In-place `map!` is even faster. It's a little strange to do bitwise/logical operations using comparisons, but it is consistent and works… and otherwise, a whole slew of new names would need to be exported (e.g., nor, nand, implies, etc.).  Also cool: this gives a very performant in-place inverse method — `map!(~, p)` — without any temporaries or new function exports.

Demo:

```julia
### BEFORE ###
julia> p = bitrand(10^8); q = bitrand(10^8); r = bitrand(10^8);

julia> map(&, p, q); @time map(&, p, q);
elapsed time: 4.875528337 seconds (12513944 bytes allocated)

julia> map!(&, r, p, q); @time map!(&, r, p, q);
elapsed time: 9.50449699 seconds (1599991904 bytes allocated, 16.76% gc time)

julia> p & q; @time p & q;
elapsed time: 0.013313429 seconds (12500176 bytes allocated)

julia> map!(~, p); @time map!(~, p);
elapsed time: 9.311276771 seconds (1599991904 bytes allocated, 17.19% gc time)

### AFTER ###
julia> map(&, p, q); @time map(&, p, q);
elapsed time: 0.016313608 seconds (12513904 bytes allocated)

julia> map!(&, r, p, q); @time map!(&, r, p, q);
elapsed time: 0.011889581 seconds (80 bytes allocated)

julia> p & q; @time p & q;
elapsed time: 0.01494325 seconds (12500176 bytes allocated)

julia> map!(~, p); @time map!(~, p);
elapsed time: 0.008248346 seconds (80 bytes allocated)

julia> f!(r,p,q) = for i=1:length(p); r[i] = p[i] & q[i]; end
       f!(r,p,q); @time f!(r, p, q); # Much faster than manual devectorization, too
elapsed time: 1.299135844 seconds (80 bytes allocated)
```

There are a few other changes here:

* The old map definitions in bitarray.jl weren't really bitarray defitions.  They were defining methods for `Union(StridedArray, BitArray)`… but `BitArray <: StridedArray`.  It seems like they belong much better in abstractarray.jl, and can be defined for all abstractarrays.

* There were also definitions for `map!` with combinations of numbers.  I think those should be deprecated in favor of `broadcast!`.

cc @carlobaldassi again - thanks!