@carlobaldassi 2015-01-13
This is great! I also much needed something like this in my own code, where I had to use BitArrays internal fields in order to avoid allocation.

It's a bit unfortunate that we can't expose all possible 2-argument functions for the lack of a corresponding name in Base.