@carlobaldassi 2015-01-13
I don't think we should deprecate the `map` versions which use `Number` arguments. I think that they follow from making scalars iterable. Maybe they should just be added in abstractarray.jl.