@mbauman 2015-01-13
I decided to deprecate `map(… ::Number …)` for the following reasons:

* As written, it only works with one- or two-argument functions.  `map!(+, zeros(2), 1, [1,2])` works but `map!(+, zeros(2), 1, [1,2], [3,4])` doesn't.

* The broadcast! function already does this much more robustly.

* Documentation for map doesn't say anything about broadcasting scalars.

But if this is at all controversial I'll happily split it out into a separate issue.