@mbauman 2015-01-14
I've refactored Functors as suggested.  See what you think.  Also now cc @lindahua since this is getting into the reduce codebase.