@carlobaldassi 2015-01-14
Travis error is perplexing but seems unrelated.
The refactoring seems good to me, let's also wait for @lindahua's opinion.

about this:
> I'm not sure it's worth taking the time to check for `one` and `zero` in specializing the functor types — there's already `fill!` that does this sort of operation very well.

it's nice for completeness, and after all even mapping the identity does not seem all that useful. I don't think the time spent in the check would be in any way significant.

Anyway, the way I see it, one of the nice things of this PR besides the non-allocating maps is that it also allows to work-around the limitations of `map` when applied to BitArrays. The reason `map` can't directly use the trick of `findnext`, i.e. evaluating the function only a bunch of times and reusing the results, is of course that functions may have state (e.g. `rand`). But if the user knows that her function does not have state, than she can map it to one of the 16 possible functions which transform 2 bits into 1 bit. In other words, one could define a `bitmap` function (disregard the silly name) which evaluates `f(true, true)`, `f(true, false)` etc. and then uses the table which you wrote in the comments. Except some functions are missing now, which is why I'd actually like to see them all implemented somehow. Anyway this doesn't need to be completed immediately, I think we can merge the PR and do that at a later time.