@mbauman 2015-01-14
The travis error looked like it stemmed from a Sets worker failing.  I've rebased onto the latest master to see if we can get around these intermittent failures.

Unfortunately, `one` and `zero` are unary operations, so they won't work to fill out the binary operators.  I can add them as `BitFunc{1}`s, though, and check for them in the unary case.

Here's a thought: What about a generic `BitFunctor` type:

```julia
immutable BitFunctor{TT,TF,FT,FF} <: Func{2} end
BitFunctor(f::Function) = is(f, &) ? BitFunctor{true, false, false false}() : …
BitFunctor(; TT=error("all truth values must be specified"), TF=error(), FT=error(), FF=error()) = BitFunctor{TT,TF,FT,FF}()
call(::BitFunctor{true, false, false, false}, x, y) = x & y
…
```

This way, even if the function doesn't exist in Base, you can still map the operation performantly.