@carlobaldassi 2015-01-14
> Unfortunately, `one` and `zero` are unary operations, so they won't work to fill out the binary operators.

I was simply thinking about this

```
immutable BinaryOneFun <: BitFunc{2} end
call(::BinaryOneFun, x, y) = _msk64

immutable BinaryZeroFun <: BitFunc{2} end
call(::ImpliesFun, x, y) = uint64(0)
```

etc. (same for unary)

> Here's a thought: What about a generic `BitFunctor` type:

I like that. Maybe avoid keyword arguments for performance: we'd just need `BitFunctor(::Function)` and

```
BitFunctor(TT,TF,FT,FF) = BitFunctor{Bool(TT),Bool(TF),Bool(FT),Bool(FF)}()`
```