@carlobaldassi 2015-01-18
@mbauman I see you have implemented the last changes - which look great. Is there anything else to do here? Tests pass and the merge conflict is trivial.

For the record, I tried the branch since I wanted to check type inference, which does not seem worse then before this PR (e.g. `map(&, ::BitArray, ::BitArray)` return type is inferred as `AbstractArray{T,N}`, and I doubt it could be better than that). BTW this is now probably the most important advantage of the generic functions over the `map` versions.