@mbauman 2015-01-18
Yup! This is good to go as far as I'm concerned.  Thanks again for all the feedback!